test: moved tests for import test and fixture, and added test for attached tests
