--- test/functional/fixtures/api/es-next/fixture/testcafe-fixtures/fixture-test.js
@@ -1,9 +1,0 @@
-import { fixture } from 'testcafe';
-
-fixture `Fixture`
-    .page `http://localhost:3000/fixtures/api/es-next/fixture/pages/index.html`;
-
-test('Fixture', async t => {
-    await t
-        .click('#button1');
-});
