+++ test/functional/fixtures/api/es-next/test-structure/testcafe-test-structure/test-structure-test.js
@@ -1,0 +1,3 @@
+// NOTE: We can just import file with tests to run them.
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import attachedTest from './attached-test';
